% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knit_with_pandoc.R
\name{knit_with_pandoc}
\alias{knit_with_pandoc}
\title{Use Pandoc Version}
\usage{
knit_with_pandoc(input, use.pandoc.version = "3.1.2", ...)
}
\arguments{
\item{input}{Default to file with focus}

\item{use.pandoc.version}{Default to '3.1.2'}

\item{...}{Other parameters}
}
\value{
Nothing but compiled file
}
\description{
Use Pandoc Version
}
